+++ Bank.XMLWebApi/Controllers/LoginEventController.cs
@@ -2,6 +2,7 @@
 using Bank.Core.Extensions;
 using Bank.Core.Utilities.XMLSerializeToXML;
 using Bank.Entity.Concrete;
+using Bank.Entity.DTOs;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -10,6 +11,8 @@ namespace Bank.XMLWebApi.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Consumes("application/xml")]
+    [Produces("application/xml")]
     public class LoginEventController : ControllerBase
     {
         private readonly ILoginEventService _loginEventService;
@@ -18,33 +21,15 @@ namespace Bank.XMLWebApi.Controllers
         {
             _loginEventService = loginEventService;
         }
-
-        [HttpGet("getall/xml")]
+        [Authorize(Roles = "Administrator")]
+        [HttpGet("getall")]
         public async Task<IActionResult> GetAllAsXml()
         {
             var result = await _loginEventService.GetAll("Time", true);
 
-            if (!result.Success)
-                return BadRequest(result);
-
-            string xmlString = XmlHelper.SerializeToXml(result.Data);
-
-
-            string xsdPath = Path.Combine(Directory.GetCurrentDirectory(), "Schemas", "LoginEvent.xsd");
-
-
-            bool isValid = XmlValidator.ValidateXml(xmlString, xsdPath, out var errors);
-
-            if (!isValid)
-            {
-                return BadRequest(new
-                {
-                    Message = "XML validation failed",
-                    Errors = errors
-                });
-            }
-
-            return Content(xmlString, "application/xml");
+            if (result.Success)
+                return Ok(result);
+            return BadRequest(result);
         }
 
 
@@ -53,8 +38,9 @@ namespace Bank.XMLWebApi.Controllers
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
             var result = await _loginEventService.GetById(id);
+            string xmlString = XmlHelper.SerializeToXml(result.Data);
             if (result.Success)
-                return Ok(result);
+                return Content(xmlString.ToString(), "application/xml");
             return BadRequest(result);
         }
 
@@ -62,6 +48,20 @@ namespace Bank.XMLWebApi.Controllers
         [HttpPost("add")]
         public async Task<IActionResult> Add([FromBody] LoginEvent loginEvent)
         {
+            string xmlString = XmlHelper.SerializeToXml(loginEvent);
+
+            string xsdPath = @"C:\Users\fb_go\source\repos\Bank.XMLWebApi\Bank.XMLWebApi\Schemas\LoginEvent.xsd";
+
+
+            bool isValid = XmlValidator.ValidateXml(xmlString, xsdPath, out var errors);
+            if (!isValid)
+            {
+                return BadRequest(new
+                {
+                    Message = "XML validation failed",
+                    Errors = errors
+                });
+            }
             var result = await _loginEventService.Add(loginEvent);
             if (result.Success)
                 return Ok(result);
@@ -72,6 +72,20 @@ namespace Bank.XMLWebApi.Controllers
         [HttpPut("update")]
         public async Task<IActionResult> Update([FromBody] LoginEvent loginEvent)
         {
+            string xmlString = XmlHelper.SerializeToXml(loginEvent);
+
+            string xsdPath = @"C:\Users\fb_go\source\repos\Bank.XMLWebApi\Bank.XMLWebApi\Schemas\LoginEvent.xsd";
+
+
+            bool isValid = XmlValidator.ValidateXml(xmlString, xsdPath, out var errors);
+            if (!isValid)
+            {
+                return BadRequest(new
+                {
+                    Message = "XML validation failed",
+                    Errors = errors
+                });
+            }
             var result = await _loginEventService.Update(loginEvent);
             if (result.Success)
                 return Ok(result);
@@ -82,6 +96,20 @@ namespace Bank.XMLWebApi.Controllers
         [HttpDelete("delete")]
         public async Task<IActionResult> Delete([FromBody] LoginEvent loginEvent)
         {
+            string xmlString = XmlHelper.SerializeToXml(loginEvent);
+
+            string xsdPath = @"C:\Users\fb_go\source\repos\Bank.XMLWebApi\Bank.XMLWebApi\Schemas\LoginEvent.xsd";
+
+
+            bool isValid = XmlValidator.ValidateXml(xmlString, xsdPath, out var errors);
+            if (!isValid)
+            {
+                return BadRequest(new
+                {
+                    Message = "XML validation failed",
+                    Errors = errors
+                });
+            }
             var result = await _loginEventService.Delete(loginEvent);
             if (result.Success)
                 return Ok(result);
